+++ pyramid_point_render_trees.cc
@@ -388,7 +388,7 @@ void PyramidPointRenderTrees::projectSurfels( Primitives* prim )
 
 double PyramidPointRenderTrees::computeHalfPixelSize( void ) {
 
-  double d = pow(2.0, cur_level) / (double)(canvas_width);
+  double d = pow(2.0, (double)cur_level) / (double)(canvas_width);
   d *= 0.5;
 
   return d;
